api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agnt-api"
version = "0.1.0"
description = "API service for the agnt marketplace"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.0",
  "greenlet>=3.0.0",
  "asyncpg>=0.29.0",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "pydantic-settings>=2.5.0",
  "aioboto3>=13.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agnt_api*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
